Extract NftDisplay and FTDisplay for reuse
