+++ web/components/Profile/NftList.tsx
@@ -1,9 +1,9 @@
 import { Text } from '@components/Common/Text';
 import { type Nft } from '@components/Swap/types';
-import { NftDisplay } from '@components/Swap/TokenSelection';
 import { FlexDiv } from '@components/Common/Alignment';
 import styled, { useTheme } from 'styled-components';
 import { spacing } from '@themes/spacing';
+import { NftDisplay } from '@components/Tokens/NftDisplay';
 
 const NftDisplaySlot = styled(FlexDiv)`
   margin-right: ${`${spacing.spacing_s}px`};
@@ -13,7 +13,7 @@ export const NftList = (props: { rawNfts: Nft[] | undefined }): JSX.Element => {
   const theme = useTheme();
 
   if (props.rawNfts === undefined) {
-    return <Text>Error while loading NFTs...</Text>;
+    return <Text>Loading NFTs...</Text>;
   }
   if (props.rawNfts.length === 0) {
     return <Text>No NFTs found</Text>;
@@ -27,6 +27,7 @@ export const NftList = (props: { rawNfts: Nft[] | undefined }): JSX.Element => {
             <NftDisplay
               nft={nft}
               key={nft.tokenId}
+              imgSize={180}
               isSelected={false}
               onClick={(e) => {}}
               captionColor={theme.properties.colorBgText}
